% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twilio_meta.R
\name{.tw_get_phones}
\alias{.tw_get_phones}
\alias{.tw_find_phones}
\title{Purchase Twilio Phones Algorithmically}
\usage{
.tw_get_phones(
  phone,
  locality = NA,
  state = NA, 
  project, 
  max_dist, 
  msgsvc_sid, 
  sid, token, 
  progress=FALSE, verbose=FALSE)
}
\arguments{
\item{phone}{Single phone number to match, 10 digit (2025555555) or 12 digit (+12025555555), or 3-digit area code. TODO: Other formats.}

\item{locality}{Optional vector of localities, as a supplement or alternative to providing phones/area codes.}

\item{state}{Optional vector of states --- if \code{state} is supplied & twilio's response phone is out of state, phone will not be purchased.}

\item{max_dist}{Maximum search distance from provided phone}

\item{sid}{Twilio account SID}

\item{token}{Twilio account Token}

\item{progress}{Show progress bar --- disabled by default when this is used outside context of \code{tw_get_phones()}.}

\item{verbose}{For error diagnosis.}

\item{project}{Project name---either "Working America" or "Mercury Opinion". Will be incorporated into phone Friendly Name.}

\item{msgsvc_sid}{The message service SID to add the phone to.}
}
\description{
Intended use: Take a random sample of phones from the list you need to text.
Use this function to actually purchase and add to an existing message service.
}
\examples{

\dontrun{
source("C:/Users/lwolberg/Desktop/.ssh/twilio_auth.R")

# NOTE: you could use the internal function that takes a single phone at a time 
# if you wanted to, for some reason.
.tw_get_phones <- function(phone, state, project, msgsvc_sid, sid, token)

}
}
\keyword{message}
\keyword{service}
\keyword{spoke}
\keyword{twilio}
