% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_messages.R
\name{get_messages}
\alias{get_messages}
\title{Get Logs of Messages Sent to Your Account}
\usage{
get_messages(page = 0, page_size = 50)
}
\arguments{
\item{page}{The page number of the log you would like to retrieve. Starts at zero.}

\item{page_size}{The number of messages per page. The maximum number allowed is 1000.}
}
\value{
A \code{twilio_message_log} object.
}
\description{
Get Logs of Messages Sent to Your Account
}
\examples{
\dontrun{

# Set API credentials
# You only need to do this once per R session
Sys.setenv(TWILIO_SID = "M9W4Ozq8BFX94w5St5hikg7UV0lPpH8e56")
Sys.setenv(TWILIO_TOKEN = "483H9lE05V0Jr362eq1814Li2N1I424t")

# Get messages sent to your account
messages <- get_messages()

}
}

