% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_tbl.R
\name{message_tbl}
\alias{message_tbl}
\title{Make a Data Frame from Message Logs}
\usage{
message_tbl(message_log)
}
\arguments{
\item{message_log}{An S3 object with the class \code{twilio_message_log}. Likely
the result of a call to \code{\link{get_messages}}.}
}
\value{
A data frame.
}
\description{
Make a Data Frame from Message Logs
}
\examples{
\dontrun{

# Set API credentials
# You only need to do this once per R session
Sys.setenv(TWILIO_SID = "M9W4Ozq8BFX94w5St5hikg7UV0lPpH8e56")
Sys.setenv(TWILIO_TOKEN = "483H9lE05V0Jr362eq1814Li2N1I424t")

# Get messages sent to your account
messages <- get_messages()

# Create data frame from log
twilio_df <- message_tbl(messages)

}
}

