% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw.auth.R
\name{tw.auth}
\alias{tw.auth}
\title{Twilio Authentication Check}
\usage{
tw.auth( sid, token )
}
\arguments{
\item{sid}{twilio credentials: SID or user key. String. Don't store this in scripts!}

\item{token}{twilio credentials: Auth token or user secret. String. Don't store this in scripts!}
}
\description{
Twilio Authentication Check
}
\section{Returns}{

Object of class "request" (\code{httr::authenticate(sid,token)})
}

\section{Permanently save twilio credentials as environment variables}{

Run this:
\code{usethis::edit_r_environ()}

It'll open a file called \code{.Renviron}, which is saved in your home folder. (On Windows, this is of the form \code{C:/Users/myusername/Documents/.Renviron}.) Enter environment variables like so (no quotes), save, close, and restart R.

\code{TWILIO_ACCOUNT_SID = ACxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}
\code{TWILIO_AUTH_TOKEN = xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx}

Now you can access environment variables using \code{Sys.getenv()}, and \code{tw.auth()} works by default.
}

\examples{

tw.auth(sid="ACxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx", token="xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx")

}
\keyword{api}
\keyword{sms}
\keyword{spoke}
\keyword{twilio}
