% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw.httr.R
\name{tw.httr}
\alias{tw.httr}
\title{Twilio Spoke API: HTTP Request Response Handling}
\usage{
tw.httr(resp, simplifyDataFrame=FALSE, debug=FALSE)
}
\arguments{
\item{resp}{A \code{httr::response()} object; returned by \code{httr::GET()}, \code{httr::DELETE()}, etc.}

\item{simplifyDataFrame}{Attempt to simplify data responses? Parameter used in \code{jsonlite::fromJSON()}.}

\item{debug}{Return original response without attempting interpretation.}
}
\description{
Twilio Spoke API: HTTP Request Response Handling
}
\keyword{api}
\keyword{sms}
\keyword{spoke}
\keyword{twilio}
