% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw.incoming_phone_numbers.R
\name{tw.incoming_phone_numbers.delete}
\alias{tw.incoming_phone_numbers.delete}
\title{Twilio Spoke API: Delete a phone from your account}
\usage{
tw.incoming_phone_numbers.delete(
  sid = Sys.getenv('TWILIO_ACCOUNT_SID'),
  key = Sys.getenv('TWILIO_KEY'),
  secret = Sys.getenv('TWILIO_SECRET'),
  phone_sid, 
  verbose = FALSE)
}
\arguments{
\item{sid}{twilio credentials: Account SID.}

\item{key}{twilio credentials: Account user key.}

\item{secret}{twilio credentials: User secret. Don't store this in scripts!!!}

\item{phone_sid}{twilio phone record SID (format 'PNXXX...XXX')}

\item{verbose}{Print detailed progress.}
}
\description{
Twilio Spoke API: Delete a phone from your account
}
\section{Details}{

?
}

\examples{

\dontrun{
source("C:/Users/lwolberg/Desktop/.ssh/twilio_auth.R")

tw.incoming_phone_numbers.delete(sid,token,'PN41fb306bf5c87631fb85ccabd3905843')
}
}
\keyword{api}
\keyword{sms}
\keyword{spoke}
\keyword{twilio}
