% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw.incoming_phone_numbers.R
\name{tw.incoming_phone_numbers.list}
\alias{tw.incoming_phone_numbers.list}
\title{Twilio Spoke API: List Owned Phones}
\usage{
tw.incoming_phone_numbers.list(
  sid = Sys.getenv('TWILIO_ACCOUNT_SID'),
  key = Sys.getenv('TWILIO_KEY'),
  secret = Sys.getenv('TWILIO_SECRET'),
  n = 20, 
  phone_number = NA, 
  friendly_name = NA, 
  verbose = FALSE
)
}
\arguments{
\item{sid}{twilio credentials: Account SID.}

\item{key}{twilio credentials: Account user key.}

\item{secret}{twilio credentials: User secret. Don't store this in scripts!!!}

\item{n}{number of records to return, for API calls that return more than one record.}

\item{phone_number}{The phone numbers of the IncomingPhoneNumber resources to read. You can specify partial numbers and use '*' as a wildcard for any digit.}

\item{friendly_name}{Provide full (not partial :P) friendly name to match only a subset of phones.}

\item{verbose}{Print detailed progress.}
}
\description{
Twilio Spoke API: List Owned Phones
}
\section{Returns}{

A \code{data.table}. NOTE: twilio API allows max 1000 to be returned.
}

\section{Details}{

\code{tw.incoming_phone_numbers.list()} gets a list of owned phone numbers.

\code{tw.incoming_phone_numbers.fetch()} gets the details of a specific owned phone number.

\code{tw.incoming_phone_numbers.delete()} deletes a specific owned phone number from your twilio account. This removes it from your bill.

\code{tw.messages.list()} gets the most recent \code{n} messages. You can specify variables to narrow the query.
}

\examples{

\dontrun{
source("C:/Users/lwolberg/Desktop/.ssh/twilio_auth.R")

tw.incoming_phone_numbers.list(sid, key, secret, n=20)
}
}
\keyword{api}
\keyword{sms}
\keyword{spoke}
\keyword{twilio}
