% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw.messages.R
\name{tw.messages.list}
\alias{tw.messages.list}
\title{Twilio Spoke API: Latest Messages}
\usage{
tw.messages.list(
  sid = Sys.getenv('TWILIO_ACCOUNT_SID'),
  key = Sys.getenv('TWILIO_KEY'),
  secret = Sys.getenv('TWILIO_SECRET'),
  from = NA, 
  to = NA, 
  datesent = NA, 
  n = 20, 
  verbose = FALSE
)
}
\arguments{
\item{sid}{twilio credentials: Account SID.}

\item{key}{twilio credentials: Account user key.}

\item{secret}{twilio credentials: User secret. Don't store this in scripts!!!}

\item{from}{number a message was sent from}

\item{to}{number a message was sent to}

\item{datesent}{date a message was sent, 'YYYY-MM-DD' format}

\item{n}{number of records to return, for API calls that return more than one record.}

\item{verbose}{Print detailed progress.}
}
\description{
Twilio Spoke API: Latest Messages
}
\section{Details}{

?
}

\examples{

\dontrun{
source("C:/Users/lwolberg/Desktop/.ssh/twilio_auth.R")

tw.messages.list(sid, token, n=20, from='19202893705')
}
}
\keyword{api}
\keyword{sms}
\keyword{spoke}
\keyword{twilio}
