% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw.messaging_services.R
\name{tw.messaging_services.create}
\alias{tw.messaging_services.create}
\title{Twilio Spoke API: Create Messaging Service}
\usage{
tw.messaging_services.create(
  sid = NA,  token = NA,
  baseurl = 'https://wkamspoke.herokuapp.com',
  friendly_name = NA,
  area_code_geomatch = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{sid}{twilio credentials: SID string. Don't store this in scripts!}

\item{token}{twilio credentials: Auth token string. Don't store this in scripts!}

\item{baseurl}{URL of your Spoke instance.}

\item{friendly_name}{Friendly name for the new service. Required.}

\item{area_code_geomatch}{Delivers message from US or Canadian Twilio phone number with a matching area code or overlay to your recipient.}

\item{verbose}{Print detailed progress.}
}
\description{
Twilio Spoke API: Create Messaging Service
}
\section{Returns}{

?
}

\section{Details}{

?
}

\examples{

\dontrun{
source("C:/Users/lwolberg/Desktop/.ssh/twilio_auth.R")

# create a service
m <- tw.messaging_services.create(sid, token, friendly_name = 'quite friendly!')
# list all services
x <- tw.messaging_services.list(sid, token, n=20)
# delete a service
tw.messaging_services.delete(sid, token, m$sid)
}
}
\keyword{api}
\keyword{sms}
\keyword{spoke}
\keyword{twilio}
