% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw.messaging_services.R
\name{tw.messaging_services.list}
\alias{tw.messaging_services.list}
\title{Twilio Spoke API: List Messaging Services}
\usage{
tw.messaging_services.list(
  sid = Sys.getenv('TWILIO_ACCOUNT_SID'),
  key = Sys.getenv('TWILIO_KEY'),
  secret = Sys.getenv('TWILIO_SECRET'),
  n = 20, 
  alldata = FALSE, 
  verbose = FALSE)
}
\arguments{
\item{sid}{twilio credentials: Account SID.}

\item{key}{twilio credentials: Account user key.}

\item{secret}{twilio credentials: User secret. Don't store this in scripts!!!}

\item{n}{number of records to return, for API calls that return more than one record.}

\item{alldata}{return allll the fields or just the marginally useful ones?}

\item{verbose}{Print detailed progress.}
}
\description{
Twilio Spoke API: List Messaging Services
}
\section{Details}{

?
}

\examples{

\dontrun{
source("C:/Users/lwolberg/Desktop/.ssh/twilio_auth.R")

# create a service
m <- tw.messaging_services.create(sid, token, friendly_name = 'quite friendly!')
# list all services
x <- tw.messaging_services.list(sid, token, n=20)
# delete a service
tw.messaging_services.delete(sid, token, m$sid)
}
}
\keyword{api}
\keyword{sms}
\keyword{spoke}
\keyword{twilio}
