% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw.messaging_services.R
\name{tw.messaging_services.update}
\alias{tw.messaging_services.update}
\title{Twilio Spoke API: Update a Messaging Service}
\usage{
tw.messaging_services.update(
  sid = NA, token = NA, 
  service_sid = NA,
  baseurl = NA,
  friendly_name = NA,
  area_code_geomatch = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{sid}{twilio credentials: SID string. Don't store this in scripts!}

\item{token}{twilio credentials: Auth token string. Don't store this in scripts!}

\item{service_sid}{twilio SID of messaging service to update}

\item{baseurl}{URL of your Spoke instance, e.g. 'https://wkamspoke.herokuapp.com'.}

\item{friendly_name}{Friendly name for the service.}

\item{area_code_geomatch}{Delivers message from US or Canadian Twilio phone number with a matching area code or overlay to your recipient.}

\item{verbose}{Print detailed progress.}
}
\description{
Twilio Spoke API: Update a Messaging Service
}
\section{Details}{

?
}

\examples{

\dontrun{
source("C:/Users/lwolberg/Desktop/.ssh/twilio_auth.R")

# create a service
m <- tw.messaging_services.create(sid, token, friendly_name = 'quite friendly!')
# update a service
tw.messaging_services.update(
  sid, token, 
  service_sid = m$sid,
  friendly_name = 'new friendly name',
  area_code_geomatch = FALSE
)
# list all services
x <- tw.messaging_services.list(sid, token, n=20)
# delete a service
tw.messaging_services.delete(sid, token, m$sid)
}
}
\keyword{api}
\keyword{sms}
\keyword{spoke}
\keyword{twilio}
