% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_buy_phone.R
\name{tw_buy_phone_number}
\alias{tw_buy_phone_number}
\title{Add a new phone number to your account}
\usage{
tw_buy_phone_number(phone_number)
}
\arguments{
\item{phone_number}{String representing phone number to purchase. See notes.}
}
\description{
To send messages with Twilio you must purchase phone numbers.
}
\note{
Phone numbers must be in E.164 format. '*' represents a wildcard digit.
}
