% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_get_calls_list.R
\name{tw_get_calls_list}
\alias{tw_get_calls_list}
\title{Get List of Calls Sent and Received from Your Account}
\usage{
tw_get_calls_list(
  page = 0,
  page_size = 50,
  status = "",
  start_time = "",
  end_time = ""
)
}
\arguments{
\item{page}{The page number of the list you would like to retrieve. Starts at zero. If `NULL`,
it will return all available data.}

\item{page_size}{The number of calls per page. The maximum number allowed is 1000.}

\item{status}{The status of the calls to include. Can be: "queued", "ringing", "in-progress",
"canceled", "completed", "failed", "busy", or "no-answer".}

\item{start_time}{Only include calls that started on this date. Specify a date as YYYY-MM-DD in
GMT, for example: `2009-07-06`, to read only calls that started on this date. You can also
specify an inequality, such as `<=YYYY-MM-DD`, to read calls that started on or before
midnight of this date, and `>=YYYY-MM-DD` to read calls that started on or after
midnight of this date.}

\item{end_time}{Only include calls that ended on this date. Specify a date as =YYYY-MM-DD in GMT,
for example: `=2009-07-06`, to read only calls that ended on this date. You can also specify
an inequality, such as `<=YYYY-MM-DD`, to read calls that ended on or before midnight of
this date, and `>=YYYY-MM-DD` to read calls that ended on or after midnight of this date.}
}
\value{
A \code{twilio_calls_list} object.
}
\description{
Retrieves a list of Twilio calls sent and received from your
account.
}
\examples{
\dontrun{
# Set API credentials
# You only need to do this once per R session
Sys.setenv(TWILIO_SID = "M9W4Ozq8BFX94w5St5hikg7UV0lPpH8e56")
Sys.setenv(TWILIO_TOKEN = "483H9lE05V0Jr362eq1814Li2N1I424t")

# Get calls sent to your account
calls <- tw_get_calls_list()
}

}
