% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_get_messages_list_by_phone.R
\name{tw_get_messages_list_by_phone}
\alias{tw_get_messages_list_by_phone}
\title{Get List of Messages Sent and Received from Your Account}
\usage{
tw_get_messages_list_by_phone(page = 0, page_size = 50, to = null)
}
\arguments{
\item{page}{The page number of the list you would like to retrieve. Starts at zero.}

\item{page_size}{The number of messages per page. The maximum number allowed is 1000.}

\item{to}{the phone number in 10-digit format you're searching for}
}
\value{
A \code{twilio_messages_list} object.
}
\description{
Retrieves a list of Twilio SMS and MMS messages sent and receieved from your
account.
}
\examples{
\dontrun{

# Set API credentials
# You only need to do this once per R session
Sys.setenv(TWILIO_SID = "M9W4Ozq8BFX94w5St5hikg7UV0lPpH8e56")
Sys.setenv(TWILIO_TOKEN = "483H9lE05V0Jr362eq1814Li2N1I424t")

# Get messages sent from your account to a phone number
messages <- tw_get_messages_list_by_phone()

}
}
