% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_lookup.R
\name{tw_lookup}
\alias{tw_lookup}
\title{tw_lookup}
\usage{
tw_lookup(num, is_valid = FALSE)
}
\arguments{
\item{num}{\code{(character)} of the number to lookup}

\item{is_valid}{\code{(logical)} \code{FALSE} (*Default*) returns the info returned by the API as a list, or \code{FALSE} for invalid numbers. \code{TRUE} will return a logical that indicates whether \code{num} is valid. Useful for subsetting.}
}
\value{
A \code{list} or \code{logical} based on the \code{is_valid} flag.
}
\description{
Fetch number info from \href{https://www.twilio.com/docs/lookup}{Twilio's number lookup service}. Useful for validating numbers.
}
