% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw.available_phone_local.R
\name{tw.available_phone_local}
\alias{tw.available_phone_local}
\title{Twilio Spoke API: Find Phone Number to Purchase}
\usage{
tw.available_phone_local(
  sid = Sys.getenv('TWILIO_ACCOUNT_SID'),
  key = Sys.getenv('TWILIO_KEY'),
  secret = Sys.getenv('TWILIO_SECRET'),
  page_size = 1,
  area_code = NA,
  near_number = NA,
  distance = ifelse(is.na(near_number),NA,25),
  in_postal_code = NA,
  in_locality = NA,
  verbose = FALSE,
  plyr = FALSE
)
}
\arguments{
\item{sid}{twilio credentials: Account SID.}

\item{key}{twilio credentials: Account user key.}

\item{secret}{twilio credentials: User secret. Don't store this in scripts!!!}

\item{page_size}{Number of results to return. Default is 1.}

\item{area_code}{The area code of the phone numbers to read. Applies to only phone numbers in the US and Canada.}

\item{near_number}{Given a phone number, find a geographically close number within distance miles. Distance defaults to 25 miles. Applies to only phone numbers in the US and Canada.}

\item{distance}{The search radius, in miles, for a near_ query. Can be up to 500 and the default is 25. Applies to only phone numbers in the US and Canada.}

\item{in_postal_code}{Limit results to a particular postal code. Given a phone number, search within the same postal code as that number. Applies to only phone numbers in the US and Canada.}

\item{in_locality}{Limit results to a particular locality or city. Given a phone number, search within the same Locality as that number.}

\item{verbose}{Print detailed progress.}

\item{plyr}{Removes a column that plyr can't deal with, to allow parallelization}
}
\description{
Twilio Spoke API: Find Phone Number to Purchase
}
\section{Returns}{

?
}

\section{Details}{

?
}

\examples{

\dontrun{
source("C:/Users/lwolberg/Desktop/.ssh/twilio_auth.R")

x <- tw.available_phone_local(sid,key,secret, area_code = 303, page_size=10)

x <- tw.available_phone_local(sid,key,secret, near_number = 7205559860, distance = 100, page_size=10)

(x <- tw.available_phone_local(sid,key,secret, in_postal_code = 81611, page_size=10))
}
}
\keyword{api}
\keyword{sms}
\keyword{spoke}
\keyword{twilio}
