% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw.incoming_phone_numbers.R
\name{tw.incoming_phone_numbers.create}
\alias{tw.incoming_phone_numbers.create}
\title{Twilio Spoke API: Purchase Phone Number}
\usage{
tw.incoming_phone_numbers.create(
  sid = NA, token = NA, 
  area_code = NA, 
  phone_number = NA,
  voice_url = '', 
  friendly_name = NA,
  verbose = FALSE )
}
\arguments{
\item{sid}{twilio credentials: SID string. Don't store this in scripts!}

\item{token}{twilio credentials: Auth token string. Don't store this in scripts!}

\item{area_code}{The desired area code for your new incoming phone number. Can be any three-digit, US or Canada area code. We will provision an available phone number within this area code for you. You must provide an area_code or a phone_number. (US and Canada only).}

\item{phone_number}{The phone number to purchase specified in E.164 format. E.164 phone numbers consist of a + followed by the country code and subscriber number without punctuation characters. For example, +14155551234.}

\item{voice_url}{The URL that we should call to answer a call to the new phone number.}

\item{friendly_name}{A descriptive string that you created to describe the new phone number. It can be up to 64 characters long. By default, this is a formatted version of the new phone number.}

\item{verbose}{Print detailed progress.}
}
\description{
Twilio Spoke API: Purchase Phone Number
}
\section{Returns}{

?
}

\section{Details}{

?
}

\examples{

\dontrun{
source("C:/Users/lwolberg/Desktop/.ssh/twilio_auth.R")

x <- tw.incoming_phone_numbers.create(
  sid, token,
  area_code = 480,
  voice_url = '',
  friendly_name = 'Working America (AZ 480)'
)
httr::content(x, "text", encoding = "UTF-8")
}
}
\keyword{api}
\keyword{sms}
\keyword{spoke}
\keyword{twilio}
