% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twilio.R
\name{tw.incoming_phone_numbers.delete}
\alias{tw.incoming_phone_numbers.delete}
\title{Twilio Spoke API: Delete a phone from your account}
\usage{
tw.incoming_phone_numbers.delete(sid = NA, token = NA, phone_sid, verbose = FALSE)
}
\arguments{
\item{sid}{twilio credentials: SID string. Don't store this in scripts!}

\item{token}{twilio credentials: Auth token string. Don't store this in scripts!}

\item{phone_sid}{twilio phone record SID (format 'PNXXX...XXX')}

\item{verbose}{Print detailed progress.}
}
\description{
Twilio Spoke API: Delete a phone from your account
}
\section{Details}{

?
}

\examples{

\dontrun{
source("C:/Users/lwolberg/Desktop/.ssh/twilio_auth.R")

tw.incoming_phone_numbers.delete(sid,token,'PN41fb306bf5c87631fb85ccabd3905843')
}
}
\keyword{api}
\keyword{sms}
\keyword{spoke}
\keyword{twilio}
