% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twilio.R
\name{tw.incoming_phone_numbers.fetch}
\alias{tw.incoming_phone_numbers.fetch}
\title{Twilio Spoke API: Fetch Details on One Phone}
\usage{
tw.incoming_phone_numbers.fetch(sid = NA, token = NA, phone_sid, verbose = FALSE)
}
\arguments{
\item{sid}{twilio credentials: SID string. Don't store this in scripts!}

\item{token}{twilio credentials: Auth token string. Don't store this in scripts!}

\item{phone_sid}{twilio phone record SID (format 'PNXXXX...XXXX')}

\item{verbose}{Print detailed progress.}
}
\description{
Twilio Spoke API: Fetch Details on One Phone
}
\section{Details}{

?
}

\examples{

\dontrun{
source("C:/Users/lwolberg/Desktop/.ssh/twilio_auth.R")

tw.incoming_phone_numbers.fetch(
  sid, token, 
  phone_sid='PN9e35ee12e6e9aaf3fc5dc39a5025df3c'
)
}

}
\keyword{api}
\keyword{sms}
\keyword{spoke}
\keyword{twilio}
