% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw.messaging_services.R
\name{tw.messaging_services.delete}
\alias{tw.messaging_services.delete}
\title{Twilio Spoke API: Delete a Messaging Service}
\usage{
tw.messaging_services.delete(
  sid = Sys.getenv('TWILIO_ACCOUNT_SID'),
  key = Sys.getenv('TWILIO_KEY'),
  secret = Sys.getenv('TWILIO_SECRET'),
  service_sid = NA, verbose = FALSE
)
}
\arguments{
\item{sid}{twilio credentials: Account SID.}

\item{key}{twilio credentials: Account user key.}

\item{secret}{twilio credentials: User secret. Don't store this in scripts!!!}

\item{service_sid}{twilio SID of messaging service to delete}

\item{verbose}{Print detailed progress.}
}
\description{
Twilio Spoke API: Delete a Messaging Service
}
\section{Returns}{

?
}

\section{Details}{

?
}

\examples{

\dontrun{
source("C:/Users/lwolberg/Desktop/.ssh/twilio_auth.R")

m <- tw.messaging_services.create(sid, token, friendly_name = 'quite friendly!')
tw.messaging_services.delete(sid, token, m$sid)
}
}
\keyword{api}
\keyword{sms}
\keyword{spoke}
\keyword{twilio}
