% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw.messaging_services.R
\name{tw.messaging_services.list}
\alias{tw.messaging_services.list}
\title{Twilio Spoke API: List Messaging Services}
\usage{
tw.messaging_services.list(sid = NA, token = NA, n = 20, alldata = FALSE, verbose = FALSE)
}
\arguments{
\item{sid}{twilio credentials: SID string. Don't store this in scripts!}

\item{token}{twilio credentials: Auth token string. Don't store this in scripts!}

\item{n}{number of records to return, for API calls that return more than one record.}

\item{alldata}{return allll the fields or just the marginally useful ones?}

\item{verbose}{Print detailed progress.}
}
\description{
Twilio Spoke API: List Messaging Services
}
\section{Details}{

?
}

\examples{

\dontrun{
source("C:/Users/lwolberg/Desktop/.ssh/twilio_auth.R")

# create a service
m <- tw.messaging_services.create(sid, token, friendly_name = 'quite friendly!')
# list all services
x <- tw.messaging_services.list(sid, token, n=20)
# delete a service
tw.messaging_services.delete(sid, token, m$sid)
}
}
\keyword{api}
\keyword{sms}
\keyword{spoke}
\keyword{twilio}
