% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_check_numbers.R
\name{tw_check_numbers}
\alias{tw_check_numbers}
\title{Twilio Check Numbers API}
\usage{
tw_check_numbers(phones, sid = NA, token = NA, test = FALSE)
}
\arguments{
\item{phones}{vector of 10-digit phone numbers, preferably as character strings.}

\item{sid}{twilio credentials: SID string. Don't store this in scripts!}

\item{token}{twilio credentials: Auth token string. Don't store this in scripts!}

\item{test}{test the progress bar}
}
\description{
Get classification of phones from Twilio as mobile/landline/voip. Costs $0.005 per call, which adds up!
}
\details{
The intermediate results are written to files in your working directory ('twilio results [date] [time].csv'), to ensure minimal loss of data if the function is interrupted.
}
\examples{

cost_limit = 1
p <- round(runif(10, min=2345678901, max=3456789012))

# source("C:/Users/lwolberg/Desktop/.ssh/twilio_auth.R")

# Using the function:
# tw_check_numbers(p, sid = 'my_sid', token = 'my_token', test = TRUE)

# some good test data for error types:
# result <- tw_check_numbers( phones = c('7209339860','5555555555','2139416728'), sid, token) )
}
\keyword{sms}
\keyword{twilio}
