% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twilio_meta.R
\name{tw_get_phones}
\alias{tw_get_phones}
\title{Purchase Twilio Phones Algorithmically}
\usage{
tw_get_phones(
  phones, 
  locality = NA,
  state = NA, 
  project, 
  msgsvc_sid, 
  sid = Sys.getenv('TWILIO_ACCOUNT_SID'),
  key = Sys.getenv('TWILIO_KEY'),
  secret = Sys.getenv('TWILIO_SECRET'),
  max_dist=75
  )
}
\arguments{
\item{phones}{Vector of phones number to match, 10 digit (2025555555) or 12 digit (+12025555555), or 3-digit area code. TODO: Other formats.}

\item{locality}{Optional vector of localities, as a supplement or alternative to providing phones/area codes.}

\item{state}{Optional vector of states --- if \code{state} is supplied & twilio's response phone is out of state, phone will not be purchased.}

\item{project}{Project name. Will be incorporated into phone Friendly Name.}

\item{msgsvc_sid}{The message service SID to add the phone to.}

\item{sid}{twilio credentials: Account SID.}

\item{key}{twilio credentials: Account user key.}

\item{secret}{twilio credentials: User secret. Don't store this in scripts!!!}

\item{max_dist}{Maximum search distance from provided phone. Default 75 miles.}
}
\description{
Intended use: Take a random sample of phones from the list you need to text.
Use this function to actually purchase and add to an existing message service.
}
\examples{

\dontrun{
source("C:/Users/lwolberg/Desktop/.ssh/twilio_auth.R")

x <- fr(filename)

# number of phones to buy
(nph <- ceiling(nrow(x)/200))

# before sample, remove low-traffic area codes
if(nchar(x$cell[1])==10) {
x[, areacode := substr(cell,1,3)]
} else if (nchar(x$cell[1])==12) {
  x[, areacode := substr(cell,3,5)]
} else stop("weird phone formatting, add an area code substring option.")
x
x \%>\%
  group_by(areacode) \%>\%
  tally \%>\%
  arrange(-n) \%>\%
  print(n=100)
(limit <- nrow(x)*.001)
# limit = 30 # if the limit is too low to make sense.
x \%>\%
  group_by(areacode) \%>\%
  tally \%>\%
  filter(n<limit) \%>\%
  as.data.table -> ac_rm
x[, sample := 1]
x[areacode \%in\% ac_rm$areacode, sample := 0]

phsamp <- sample(x[sample==1, cell], size=nph, replace=F)

y <- tw_get_phones(
  phones = phsamp,
  state = NA,
  project = "Working America",
  msgsvc_sid = "MGaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
  sid=sid, key=key, secret=secret
)

}
}
\keyword{message}
\keyword{service}
\keyword{spoke}
\keyword{twilio}
