% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_lookup.R
\name{tw_lookup}
\alias{tw_lookup}
\title{Twilio Lookup}
\usage{
tw_lookup(number, is_valid = FALSE, country_code = "US")
}
\arguments{
\item{number}{A phone number to lookup.}

\item{is_valid}{Return raw API output or logical that indicates whether number is valid (default).}

\item{country_code}{The ISO country code of the phone number to fetch. This is used to specify the country when the phone number is provided in a national format.}
}
\value{
A \code{list} or \code{logical} based on the \code{is_valid} flag.
}
\description{
Fetch number info from Twilio's number lookup service. Useful for validating numbers.
}
\examples{
\dontrun{

# Set API credentials
# You only need to do this once per R session
Sys.setenv(TWILIO_SID = "M9W4Ozq8BFX94w5St5hikg7UV0lPpH8e56")
Sys.setenv(TWILIO_TOKEN = "483H9lE05V0Jr362eq1814Li2N1I424t")

# Get messages sent to your account
tw_lookup("+41757105871")

}
}
